refactor: extract request body component and headers-related actions
